% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Arthabund.R
\docType{data}
\name{Arthabund}
\alias{Arthabund}
\title{Arthropod Abundance Data}
\format{
An object of class \code{data.frame} with 10 rows and 79 columns.
}
\usage{
Arthabund
}
\description{
Arthropod abundance data from 10 sites for 78 taxonomic groups.
}
\references{
Myers, AL, and JM Marshall. 2021. Influence of forest fragment composition and structure on ground-dwelling arthropod communities. American Midland Naturalist 186:76-94.
}
\keyword{datasets}
