% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOL54210-package.R
\docType{data}
\name{Arthabund_wEnviron}
\alias{Arthabund_wEnviron}
\title{Arthropod Abundance with Environmental Data}
\format{
An object of class \code{data.frame} with 10 rows and 84 columns.
}
\usage{
Arthabund_wEnviron
}
\description{
Arthropod abundance data from 10 sites for 78 taxonomic groups with percent slope (slope), percent canopy cover (canopy), fine woody debris abundance (FWD), litter depth (cm; litter), and coarse woody debris volume (m²/ha; CWD).
}
\references{
Myers, AL, and JM Marshall. 2021. Influence of forest fragment composition and structure on ground-dwelling arthropod communities. American Midland Naturalist 186:76-94.
}
\keyword{datasets}
