% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barkRoughness.R
\docType{data}
\name{barkRoughness}
\alias{barkRoughness}
\title{Bark Roughness Data}
\format{
An object of class \code{data.frame} with 382 rows and 3 columns.
}
\usage{
barkRoughness
}
\description{
Bark roughness data (as a percent) derived from binomial images with categorical dead (1 dead, 0 alive) and diameter at breast height (cm; DBH).
}
\references{
Boyes, KN, KG Hietala-Henschell, AP Barton, AJ Storer, and JM Marshall. 2019. Linking tree growth rate, damage repair, and susceptibility to a genus-specific pest infestation. Journal of Forestry Research 30:1935-1941.
}
\keyword{datasets}
