% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOL54210-package.R
\docType{package}
\name{BIOL54210-package}
\alias{BIOL54210}
\alias{BIOL54210-package}
\title{BIOL 54210 Biometry Data}
\description{
This package contains class data for BIOL 54210 Biometry at Purdue University Fort Wayne.
}
\details{
Due to the data structure, when you load the \code{BIOL54210} package, only tabular data is automatically available. To load the spatial vector data, you need to use the \link{load_spatial} function, which will load data from the package into the Global Environment.

For the homework assignments, exams, and your own data outside of class, you will need to load external data files into R. There are several different ways to do that depending of the file format. Examples below are reminders of how to do this.
}
\examples{
# Load most text file formats - example is *.dat
data_name <- read.table("file_name.dat", header = TRUE)

# Load *.csv file format
data_name <- read.csv("file_name.csv")

# Load *.xlsx file format in R Studio
data_name <- readxl::read_excel("file_name.xlsx")

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PlantEcology/BIOL54210}
}

}
\author{
\strong{Maintainer}: Jordan Marshall \email{marshalj@pfw.edu}

}
\keyword{internal}
