% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOL54210-package.R
\docType{data}
\name{categReg}
\alias{categReg}
\title{Categorical Regression Data}
\format{
An object of class \code{data.frame} with 151 rows and 7 columns.
}
\usage{
categReg
}
\description{
Emerald ash borer capture rates (beetles/trap) at sites in Indiana, Michigan, Ohio, and Wisconsin with tree data where traps were hung: vigor (categorical, 1-6 where 1 is a healthy tree and 6 is over half dead), dieback (percent; 5\% categories), crown light exposure (CLE; categorical; number of sides exposed to full-sun; 0 fully suppressed, 5 fully open grown with four sides and top in full sun), and dominance class (1 is dominant and 5 suppressed).
}
\references{
Marshall, JM, JA Beachy, I Fraser, AJ Storer, and VC Mastro. 2011. Multistate comparison of emerald ash borer detection tools: A five year synthesis. Proceedings of 22nd U.S. Department of Agriculture Interagency Research Forum on Invasive Species, Annapolis, Maryland.
}
\keyword{datasets}
