% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closSub.R
\docType{data}
\name{closSub}
\alias{closSub}
\title{Wound Closure Data}
\format{
An object of class \code{data.frame} with 11 rows and 2 columns.
}
\usage{
closSub
}
\description{
Subset data (n = 11) of rtifical wound closure in ash (Fraxinus sp.) trees with presence of emerald ash borer exit holes (EH; 1 present and 0 absent). The full data set is closFull.
}
\references{
Boyes, KN, KG Hietala-Henschell, AP Barton, AJ Storer, and JM Marshall. 2019. Linking tree growth rate, damage repair, and susceptibility to a genus-specific pest infestation. Journal of Forestry Research 30:1935-1941.
}
\keyword{datasets}
