% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOL54210-package.R
\docType{data}
\name{compDat}
\alias{compDat}
\title{Plant Survey Data}
\format{
An object of class \code{data.frame} with 145 rows and 6 columns.
}
\usage{
compDat
}
\description{
Plant community composition data from plots along six transects with diversity (Shannon entropy), native species abundance, non-native species abundance, abundance of Centaurea stoebe (CESTM; non-native), and abundance of Stellaria longipes (STOL2; native).
}
\references{
Marshall, JM. 2014. Influence of topography, bare sand, and soil pH on the occurrence and distribution of plant species in a lacustrine dune ecosystem. Journal of the Torrey Botanical Society 141:29-38.
}
\keyword{datasets}
