% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environ.R
\docType{data}
\name{environ}
\alias{environ}
\title{Environmental Data}
\format{
An object of class \code{data.frame} with 10 rows and 6 columns.
}
\usage{
environ
}
\description{
Percent slope (slope), percent canopy cover (canopy), fine woody debris abundance (FWD), litter depth (cm; litter), and coarse woody debris volume (m²/ha; CWD) associated with Arthabund_wEnviron is pooling of this data and Arthabund.
}
\references{
Myers, AL, and JM Marshall. 2021. Influence of forest fragment composition and structure on ground-dwelling arthropod communities. American Midland Naturalist 186:76-94.
}
\keyword{datasets}
