% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exam1A.Data.R
\docType{data}
\name{Exam1A.Data}
\alias{Exam1A.Data}
\title{Exam 1 Part A Data}
\format{
An object of class \code{data.frame} with 80 rows and 6 columns.
}
\usage{
Exam1A.Data
}
\description{
Data to be used in Exam 1 Part A. Arthropod abundance with percent forest cover, percent understory cover, ratio of non-native plants (1 is all plants in understory were non-native and 0 is all plants were native) at sites in four counties and 20 plots per site.
}
\references{
Marshall, JM, and DS Buckley. 2009. Influence of Microstegium vimineum presence on insect abundance in hardwood forests. Southeastern Naturalist 8:515-526.
}
\keyword{datasets}
