% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOL54210-package.R
\docType{data}
\name{forest}
\alias{forest}
\title{Forest Spatial and Diversity Data}
\format{
An object of class \code{data.frame} with 30 rows and 16 columns.
}
\usage{
forest
}
\description{
Overstory, midstory, and understory diversity and richness at 30 sites in Adams, Wells, and Allen Counties, Indiana, with floristic quality index (FQI), percent canopy cover (Canopy), percent volumetric soil moisture content (VMC), estimated forest age (from dendrochronology), basal area (m²/ha), forest area (ha), perimeter (m), shape (perimeter:area ratio), and harvest historical occurrance (binomial). Understory diversity and richness only include woody plant species.
}
\references{
Harman, RR, AR De Leon, HM Lancaster, and JM Marshall. 2019. Relative influence of spatial and structural characteristics of forest fragments on woody plant communities. The Great Lakes Botanist 58:32-54.
}
\keyword{datasets}
