% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestEnv.R
\docType{data}
\name{forestEnv}
\alias{forestEnv}
\title{Forest Environmental Data}
\format{
An object of class \code{data.frame} with 57 rows and 6 columns.
}
\usage{
forestEnv
}
\description{
Percent canopy cover (Canopy.Cover) and litter depth (cm; Litter) in three forests in Allen County, Indiana, with an arbitrary random binary variable.
}
\references{
Arvola, KD, PF Booth, CC Ellinwood, AJ Fry, JL Furge, KA Haehnle, LE Hall, MA Jablonski, DK Jones, JT Martin, KM McLane, KC Mensch, VA Mumaw, RN Quirindongo, MJ Ravesi, JJ Rinard, PR Selig, AP Sellan, MB Sljivar, EA Stulik, TR Sunday, AN Turley, MT Voors, AR Warrix, TC Wood, and JM Marshall. 2014. Comparative analysis of urban and rural forest fragment structure and diversity in Norheastern Indiana. The Michigan Botanist 53:39-59.
}
\keyword{datasets}
