% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOL54210-package.R
\docType{data}
\name{INcovid19}
\alias{INcovid19}
\title{Covid 19 Vaccination Rates}
\format{
An object of class \code{data.frame} with 877 rows and 2 columns.
}
\usage{
INcovid19
}
\description{
Total vaccine doses administered by date for Indiana (original data by county, this is summed for the state).
}
\references{
https://hub.mph.in.gov/dataset/covid-19-vaccinations-by-date
}
\keyword{datasets}
