% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spatial.R
\name{load_spatial}
\alias{load_spatial}
\title{Load Spatial Data}
\usage{
load_spatial(whichData = "all")
}
\value{
Vector spatial data as global variables for class
}
\description{
This function loads vector spatial data needed for BIOL 54210 Biometry.
}
