% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOL54210-package.R
\docType{data}
\name{mixed.Data}
\alias{mixed.Data}
\title{Dispersal Pattern Data}
\format{
An object of class \code{data.frame} with 46 rows and 7 columns.
}
\usage{
mixed.Data
}
\description{
Dispersal distances (cm) for Microstegium vimineum in different disturbance treatments of control (C), soil disturbance (litter intact; L), litter removed (soil intact; LR), and soil disturbance with litter removal (LRL). Distances measured at the point 90\% of individuals were captured, distance 100\% of individuals captured, and the mean distance for the plot. Each treatment replicated 3 times in 4 forests.
}
\references{
Marshall, JM, and DS Buckley. 2008. Influence of litter removal and mineral soil disturbance on the spread of an invasive grass in a Central Hardwood forest. Biological Invasions 10:531-538.
}
\keyword{datasets}
