% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOL54210-package.R
\docType{data}
\name{occupancy.Data}
\alias{occupancy.Data}
\title{Presence and Absence Trapping Data}
\format{
An object of class \code{data.frame} with 189 rows and 10 columns.
}
\usage{
occupancy.Data
}
\description{
Emerald ash borer capture rates (beetles/trap) at sites in Indiana, Michigan, Missouri, Ohio, Pennsylvania, and Wisconsin using 5 different trap types.
}
\references{
Marshall, JM, JA Beachy, I Fraser, AJ Storer, and VC Mastro. 2011. Multistate comparison of emerald ash borer detection tools: A five year synthesis. Proceedings of 22nd U.S. Department of Agriculture Interagency Research Forum on Invasive Species, Annapo
}
\keyword{datasets}
