% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOL54210-package.R
\docType{data}
\name{paired.Data}
\alias{paired.Data}
\title{Paired Capture Rate Data}
\format{
An object of class \code{data.frame} with 44 rows and 4 columns.
}
\usage{
paired.Data
}
\description{
Emerald ash borer capture rates (beetles/trap) at sites in Indiana, Michigan, and Ohio in 2007 and 2017.
}
\references{
Marshall, JM, AJ Storer, I Fraser, JA Beachy, and VC Mastro. 2009. Effectiveness of differing trap types for the detection of emerald ash borer (Coleoptera: Buprestidae). Environmental Entomology 38:1226-1234.
}
\keyword{datasets}
