% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOL54210-package.R
\docType{data}
\name{repeatCap}
\alias{repeatCap}
\title{Repeated Capture Data}
\format{
An object of class \code{data.frame} with 258 rows and 4 columns.
}
\usage{
repeatCap
}
\description{
Repeated measure data for captures of emerald ash borer at sites in 6 states with 6 repeated measures.
}
\references{
Marshall, JM, JA Beachy, I Fraser, AJ Storer, and VC Mastro. 2011. Multistate comparison of emerald ash borer detection tools: A five year synthesis. Proceedings of 22nd U.S. Department of Agriculture Interagency Research Forum on Invasive Species, Annapo
}
\keyword{datasets}
