% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpModDev.R
\docType{data}
\name{rpModDev}
\alias{rpModDev}
\title{Tree Mortality Data}
\format{
An object of class \code{data.frame} with 146 rows and 6 columns.
}
\usage{
rpModDev
}
\description{
Tree mortality (binomial; 1 dead and 0 alive) for trees with diameter at breast height (DBH; cm), bark splits (BS; binomial presnce and absence), dieback (percent crown dieback 5\% categories), and vigor (categorical, 1-6 where 1 is a health tree and 6 is over half dead).
}
\references{
Hale, CE, MA Jordan, G Iriarte, K Broders, AJ Storer, VJ Nalam, and JM Marshall. 2021. Genome-wide SNP identification in Fraxinus linking genetic characteristics to tolerance of Agrilus planipennis. Ecology and Evolution 11:14775-14788.
}
\keyword{datasets}
