% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpModTest.R
\docType{data}
\name{rpModTest}
\alias{rpModTest}
\title{Tree Mortality Data}
\format{
An object of class \code{data.frame} with 148 rows and 6 columns.
}
\usage{
rpModTest
}
\description{
Same data organization as rpModDev, but a separate set of trees to test models produced from rpModDev.
}
\references{
Hale, CE, MA Jordan, G Iriarte, K Broders, AJ Storer, VJ Nalam, and JM Marshall. 2021. Genome-wide SNP identification in Fraxinus linking genetic characteristics to tolerance of Agrilus planipennis. Ecology and Evolution 11:14775-14788.
}
\keyword{datasets}
