% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOL54210-package.R
\docType{data}
\name{slopeComp2}
\alias{slopeComp2}
\title{Tree Ring Area Data}
\format{
An object of class \code{data.frame} with 47 rows and 3 columns.
}
\usage{
slopeComp2
}
\description{
Same data as GAM.Data and slopeComp, but reorganized and only Low and High. Tree ring surface area (mm²) by year for ash trees with low and high tolerance to emerald ash borer.
}
\references{
Boyes, KN, KG Hietala-Henschell, AP Barton, AJ Storer, and JM Marshall. 2019. Linking tree growth rate, damage repair, and susceptibility to a genus-specific pest infestation. Journal of Forestry Research 30:1935-1941.
}
\keyword{datasets}
