% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIOL54210-package.R
\docType{data}
\name{treeGrowth}
\alias{treeGrowth}
\title{Tree Ring Area with Climate Data}
\format{
An object of class \code{data.frame} with 25 rows and 6 columns.
}
\usage{
treeGrowth
}
\description{
Same data as GAM.Data, but with growing season mean precipitation (cm) and temperature (°C). Tree ring surface area (mm²) by year for ash trees with low, middle, and high tolerance to emerald ash borer.
}
\references{
Boyes, KN, KG Hietala-Henschell, AP Barton, AJ Storer, and JM Marshall. 2019. Linking tree growth rate, damage repair, and susceptibility to a genus-specific pest infestation. Journal of Forestry Research 30:1935-1941.
}
\keyword{datasets}
