% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeHealth.R
\docType{data}
\name{treeHealth}
\alias{treeHealth}
\title{Tree Categorical Health Data}
\format{
An object of class \code{data.frame} with 51 rows and 4 columns.
}
\usage{
treeHealth
}
\description{
Tree health (leaf color) for ash trees at sites in three states with diameter at breast height (cm).
}
\references{
Marshall, JM, AJ Storer, I Fraser, JA Beachy, and VC Mastro. 2009. Effectiveness of differing trap types for the detection of emerald ash borer (Coleoptera: Buprestidae). Environmental Entomology 38:1226-1234.
}
\keyword{datasets}
