% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spatial.R
\name{load_spatial}
\alias{load_spatial}
\title{Load Spatial Data}
\usage{
load_spatial(whichData = "all")
}
\arguments{
\item{whichData}{Specific data file needed. Options include "vector", "raster", "all". Default is all.}
}
\value{
A spatial data file for class
}
\description{
This function loads spatial data (vector and raster) needed for BIOL 54210 Biometry.
}
